server/requirements.txt
# Agri-Lo Backend Dependencies
paho-mqtt>=1.6.1

# Core Framework
fastapi
uvicorn[standard]
python-multipart

# Database
sqlmodel
aiosqlite

# Authentication
python-jose[cryptography]
passlib[argon2]

# Configuration
pydantic-settings
email-validator
python-dotenv

# AI/ML - Relaxed to avoid conflicts on Python 3.13
tensorflow
tensorflow-model-optimization
scikit-learn
joblib
tf-keras

# Data Processing
pandas
numpy
Pillow

# AI Chatbot
google-generativeai
groq

# HTTP Client
httpx

# Payments
razorpay

# Authentication (Firebase)
firebase-admin
slowapi